+++ CHANGELOG.md
@@ -58,6 +58,7 @@ high state of flux, you're at risk of it changing without notice.
   - `ReadonlySet`
     - make `isSubset` data-last (@gcanti)
     - make `elem` data-last (@gcanti)
+    - make `union` data-last (@gcanti)
   - `Record`
     - make `isSubrecord` data-last (@gcanti)
     - make `lookup` data-last (@gcanti)
@@ -65,6 +66,7 @@ high state of flux, you're at risk of it changing without notice.
   - `Set`
     - make `subset` data-last (@gcanti)
     - make `elem` data-last (@gcanti)
+    - make `union` data-last (@gcanti)
   - `Semigroup`
     - make `fold` data-last (@gcanti)